+++ Cargo.toml
@@ -1,35 +1,23 @@
-[package]
-name = "solagent"
-version = "0.1.12"
+[workspace]
+resolver = "2"
+members = [
+    "packages/core",
+    "packages/adapters/rig/goplus",
+    "packages/adapters/rig/solana",
+    "packages/adapters/topkio",
+    "packages/plugins/goplus",
+    "packages/plugins/solana", "examples/plugin",
+]
+
+[workspace.package]
+version = "0.1.0"
 edition = "2021"
-authors = ["zTgx <beautifularea@gmail.com>"]
-repository = "https://github.com/zTgx/solagent.rs.git"
-readme = "README.md"
-keywords = ["ai", "solana", "agent", "kit"]
-license = "MIT"
-description = "connect any ai agents to solana protocols in Rust."
 
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-lazy_static = "1.5.0"
-rig-core = "0.6.1"
-async-trait = "0.1.83"
-solana-client = "2.1.7"
-solana-sdk = "2.1.7"
-mpl-token-metadata = { version = "5.1.0", features = ["serde"] }
-spl-token = "7.0.0"
-spl-token-2022 = "6.0.0"
-spl-associated-token-account = "6.0.0"
-solana-program = "2.1.7"
-thiserror = "2.0.9"
-reqwest = { version = "0.12", features = ["json", "multipart"] }
-bincode = "1.3.3"
-base64 = "0.22.1"
-anyhow = "1.0.80"
-solana-account-decoder = "2.1.9"
-helius = "0.2.4"
+[profile.dev]
+opt-level = "z"
 
-[dev-dependencies]
-tokio = { version = "1.42.0", features = ["full"] }
-rig-core = "0.6.1"
+[profile.release]
+lto = "thin"
+strip = true
+codegen-units = 1
+opt-level = "z"
