+++ examples/plugin/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "plugin"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+solagent-core = { path = "../../packages/core" }
+tokio = { version = "1.42.0", features = ["full"] }
+solagent-plugin-goplus = { path = "../../packages/plugins/goplus" }
+solagent-plugin-solana = { path = "../../packages/plugins/solana" }
+solagent-rig-goplus = { path = "../../packages/adapters/rig/goplus" }
