+++ packages/adapters/rig/solana/src/close_empty_token_accounts.rs
@@ -12,16 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-    actions::{close_empty_token_accounts, CloseEmptyTokenAccountsData},
-    SolanaAgentKit,
-};
-use rig::{
-    completion::ToolDefinition,
-    tool::{Tool, ToolEmbedding},
-};
-use serde::{Deserialize, Serialize};
+use solagent_core::{rig::{completion::ToolDefinition, tool::{Tool, ToolEmbedding}}, SolanaAgentKit};
 use std::sync::Arc;
+use serde::{Deserialize, Serialize};
+use solagent_plugin_solana::{close_empty_token_accounts, CloseEmptyTokenAccountsData};
 
 #[derive(Debug, Deserialize)]
 pub struct CloseEmptyTokenAccountsArgs {}
