+++ packages/adapters/rig/solana/src/lib.rs
@@ -12,17 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod close_empty_token_accounts;
-pub use close_empty_token_accounts::{close_empty_token_accounts, CloseEmptyTokenAccountsData};
-
-mod get_balance;
-pub use get_balance::get_balance;
-
-mod request_faucet_funds;
-pub use request_faucet_funds::request_faucet_funds;
-
-mod get_tps;
-pub use get_tps::get_tps;
-
-mod transfer;
-pub use transfer::transfer;
+pub mod close_empty_token_accounts;
+pub mod get_balance;
+pub mod get_tps;
+pub mod request_faucet_funds;
+pub mod transfer;
