+++ packages/adapters/topkio/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "topkio"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
