+++ src/lib.rs
@@ -81,45 +81,3 @@ pub use spl_associated_token_account;
 pub use spl_token;
 pub use spl_token_2022;
 pub use thiserror;
-
-/// Wallet
-/// - wallet : Wallet keypair for signing transactions
-/// - address: Public key of the wallet
-struct Wallet {
-    pub(crate) wallet: Keypair,
-    pub(crate) address: Pubkey,
-}
-
-impl Wallet {
-    pub fn load(private_key: &str) -> Wallet {
-        let secret_key = bs58::decode(private_key).into_vec().expect("private key is not valid base58 format!");
-        let wallet = Keypair::from_bytes(&secret_key).expect("Invalid private key!");
-        let address = wallet.pubkey();
-
-        Wallet { wallet, address }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct Config {
-    pub openai_api_key: Option<String>,
-    pub jupiter_referral_account: Option<String>,
-    pub jupiter_fee_bps: Option<u16>, // Assuming fee is represented as a percentage (0-10000)
-    pub flash_privilege: Option<String>,
-    pub flexlend_api_key: Option<String>,
-    pub helius_api_key: Option<String>,
-}
-
-/// Represents a Solana agent that interacts with the blockchain.
-/// Provides a unified interface for token operations, NFT management, trading and more
-pub struct SolanaAgentKit {
-    wallet: Wallet,
-    config: Config,
-    connection: RpcClient,
-}
-
-impl SolanaAgentKit {
-    pub fn new(private_key: &str, rpc_url: &str, config: Config) -> Self {
-        SolanaAgentKit { wallet: Wallet::load(private_key), config, connection: RpcClient::new(rpc_url) }
-    }
-}
