--- src/tools/goplus/solana_token_security_info.rs
@@ -1,80 +1,0 @@
-// Copyright 2025 zTgx
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use crate::{actions::get_solana_token_security_info, parameters_json_schema};
-use rig::{completion::ToolDefinition, tool::Tool};
-use serde::{Deserialize, Serialize};
-use serde_json::json;
-
-#[derive(Debug, Deserialize)]
-pub struct SolanaTokenSecurityInfoArgs {
-    contract_address: String,
-}
-
-#[derive(Deserialize, Serialize)]
-pub struct SolanaTokenSecurityInfoOutput {
-    pub data: serde_json::Value,
-}
-
-#[derive(Debug, thiserror::Error)]
-#[error("SolanaTokenSecurityInfo error")]
-pub struct SolanaTokenSecurityInfoError;
-
-#[derive(Default)]
-pub struct SolanaTokenSecurityInfo;
-impl SolanaTokenSecurityInfo {
-    pub fn new() -> Self {
-        SolanaTokenSecurityInfo {}
-    }
-}
-
-impl Tool for SolanaTokenSecurityInfo {
-    const NAME: &'static str = "get_solana_token_security_info";
-
-    type Error = SolanaTokenSecurityInfoError;
-    type Args = SolanaTokenSecurityInfoArgs;
-    type Output = SolanaTokenSecurityInfoOutput;
-
-    async fn definition(&self, _prompt: String) -> ToolDefinition {
-        ToolDefinition {
-            name: "get_solana_token_security_info".to_string(),
-            description: r#"
-            
-            Token Security API for Solana (Beta).
-
-            examples: [
-                [
-                    {
-                        input: {
-                            contract_address: "So11111111111111111111111111111111111111112",
-                        },
-                    },
-                ],
-            ]
-              "#
-            .to_string(),
-            parameters: parameters_json_schema!(
-                chain_id: String,
-                contract_address: String
-            ),
-        }
-    }
-
-    async fn call(&self, args: Self::Args) -> Result<Self::Output, Self::Error> {
-        let data =
-            get_solana_token_security_info(&args.contract_address).await.expect("get_solana_token_security_info");
-
-        Ok(SolanaTokenSecurityInfoOutput { data })
-    }
-}
